Added support for list values in Value (Dust now supports nested lists)
- Updated Dust variant to include std::vector<std::shared_ptr<Value>>
- Enables recursive, expressive list structures in the language

📖 "Two are better than one, because they have a good return for their labor." — Ecclesiastes 4:9
